refactor(stories): move stories to .storybook folder
