+++ packages/components/.storybook/preview.tsx
@@ -1,6 +1,5 @@
 import "@chanzuckerberg/eds-tokens/css/variables.css";
 import "../src/styles/fonts.css";
-import "../src/styles/tailwind-variables.css";
 import "../src/styles/global.css";
 
 import * as React from "react";
