+++ packages/components/src/Button/button.stories.tsx
@@ -1,15 +1,11 @@
 import * as React from "react";
 
-import Button, { ButtonProps } from "./button";
-import Clickable, { ClickableProps } from "../util/clickable";
-
-import Heading from "../Heading";
+import Button from "./button";
 import { Story } from "@storybook/react/types-6-0";
 import Text from "../Text";
-import styles from "./button.stories.module.css";
 
 export default {
-  title: "Button",
+  title: "Button - WIP",
   component: Button,
   argTypes: {
     children: {
@@ -26,8 +22,8 @@ const Template: Story<Args> = (args) => <Button {...args} />;
 export const flat = Template.bind(null);
 flat.args = {
   children: "Flat Button",
-  variant: "flat",
   color: "success",
+  variant: "flat",
 };
 
 export const outline = Template.bind(null);
@@ -39,8 +35,8 @@ outline.args = {
 export const link = Template.bind(null);
 link.args = {
   children: "Link Button",
-  variant: "link",
   color: "alert",
+  variant: "link",
 };
 
 export const linkInBody = (args) => (
@@ -51,8 +47,8 @@ export const linkInBody = (args) => (
 );
 linkInBody.args = {
   children: "Link Button",
-  variant: "link",
   color: "brand",
+  variant: "link",
 };
 
 export const linkInHeading = (args) => (
@@ -65,72 +61,4 @@ linkInHeading.args = {
   children: "Link Button",
   variant: "link",
   color: "brand",
-};
-
-const sizes: Array<ButtonProps["size"]> = ["small", "medium", "large"];
-const colors: Array<ButtonProps["color"]> = [
-  "alert",
-  "brand",
-  "neutral",
-  "success",
-];
-const variants: Array<ButtonProps["variant"]> = ["flat", "outline", "link"];
-const states: Array<ClickableProps<"button">["state"]> = [
-  "inactive",
-  "hover",
-  "focus",
-  "active",
-  "disabled",
-];
-
-export const allVariants = () => {
-  return (
-    <ul>
-      {sizes.map((size) => (
-        <li key={size}>
-          {variants.map((variant) => (
-            <React.Fragment key={variant}>
-              <Heading size="h2">
-                {variant} - {size}
-              </Heading>
-              <table className={styles["variant"]}>
-                <tbody>
-                  {states.map((state) => (
-                    <tr key={state}>
-                      <th scope="row">
-                        <Text size="body">{state}</Text>
-                      </th>
-                      {colors.map((color) => (
-                        <td key={color} className={styles["color"]}>
-                          <Clickable
-                            as="button"
-                            size={size}
-                            color={color}
-                            variant={variant}
-                            state={state}
-                            disabled={state === "disabled"}
-                          >
-                            Button
-                          </Clickable>
-                        </td>
-                      ))}
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </React.Fragment>
-          ))}
-        </li>
-      ))}
-    </ul>
-  );
-};
-
-allVariants.parameters = {
-  snapshot: {
-    skip: true,
-  },
-  axe: {
-    skip: true,
-  },
 };
