+++ packages/components/src/Button/index.ts
@@ -1,0 +1,1 @@
+export { default, ButtonProps } from "./button";
