+++ packages/components/src/styles/global.css
@@ -1,1 +1,16 @@
+
+/* Copied from output of `@tailwind utilities` */
+* {
+  --tw-shadow: 0 0 #0000;
+  --tw-ring-inset: var(--tw-empty,/*!*/ /*!*/);
+
+  /* Can't remove the px because it breaks tailwind rings */
+  /* stylelint-disable-next-line length-zero-no-unit */
+  --tw-ring-offset-width: 0px;
+  --tw-ring-offset-color: #fff;
+  --tw-ring-color: rgba(59, 130, 246, 0.5);
+  --tw-ring-offset-shadow: 0 0 #0000;
+  --tw-ring-shadow: 0 0 #0000;
+}
+
 @tailwind base;
